Workswitch toggles auto, moved logic to class
Summary of alterations:
- Workswitch toggles auto instead of manual
- Workswitch logic extracted to seperate class/file
- Altered logic to be conditional
    i.e. The logic for toggling auto on/off only executes if/when the
         state of the workswitch changes
- Enabling the workswitch in the vehicle settings no longer disables
  the on-screen button
    - Cycling the workswitch will re-sync the switch and on-screen
      button to the same state (ensuring that the workswitch
      can't become inverted)

The logic regarding the workswitch has been relocated from
"OpenGL.Designer.cs" to "CWorkSwitch.cs"; where the logic in the
former resided, a method call to the latter has been implemented.
The neccessary instantiation and assignment required has been included
in "FormGPS.cs".
The toggling of the auto mode based on the work switch is undertaken
by simulating a "click" (press) of the on-screen button.

The location of the "CheckWorkSwitch()" method call in
"OpenGL.Designer.cs" is functional, but could be easily changed if a
more suitable location is evaluated. Checking of the workswitch is only
undertaken when workswitch is enabled and a job/field has been started.