Work switch manual section control
Work switch now has the functionality to control sections using manual
mode. This is to amend the limitation of my previous changes that only
allowed for the use of automatic section control with the work switch.

A checkbox has been added to the work switch tab of the vehicle settings
to facilitate this functionality.