Implement method to check and alter angles
Check the value that is passed in, alter (if required) and return to
caller. Use to ensure an angle is greater than or equal to zero
(not negative) and less than or equal to TAU (2 * PI), in terms of
radians.

This check is useful to ensure "heading" is between the range specified
above. For example, to correct the possible negative value that can be
returned from the trigonometry functions in "System.Math": Atan, Atan2,
Asin, Acos;

The logic used can be found duplicated throughout the code when it is
needed. Moving the logic into a method will increase code readability
and is less error prone. This method should be used instead.