Add structs to contain external values and flags
Add 'Inputs' struct to group the necessary values for u-turn
configuration and generation together. Allowing for the contained values
to be easily passed to methods, instead of passing each variable
individually, or worse, calling the external variables from within the
scope of the method.

Add 'Flags' struct to contain the const values that will be used to as
bitflags for setting and obtaining configuration information.