Implement mathematic geometric vectors
Add a type to represent a geometric vector in two dimensions. These
vectors can be declared using either: (cartesian) 'change of x' and
'change of y' or (polar) 'magnitude' and 'direction'.

The Vector2 constructor can not be called directly, instead, two methods
are provided by the Vector2 type: 'Vector2.Cartesian' and
'Vector.Polar'. Given appropriate inputs these methods do the necessary
calculations and pass the resulting values to the Vector2 constructor,
which in turn is returned to the caller.

The dual defintion removes the need to convert between coordinate
systems and the need to rememeber which system a particular vector was
defined in terms of.