Implement linear line equation as type
Used to store the values of linear equations in two dimensions,
expressed in terms of cartesian coordinates, and defined in terms of:
slope, x intercept and y intercept. Includes helper methods for defining
an equation that is parallel to a coordinate axis:

  * "DefineByX" for an equation parallel to the 'y axis'
  * "DefineByY" for an equation parallel to the 'x axis'

The appropriate slope and opposing intercept values for each situation
are implemented into each method. A line equation that is parallel to
the 'x axis' will have a 'slope' of 0, and an undefined 'x intercept'
represented using 'double.NaN'. A line equation parallel to the 'y axis'
will have an undefined 'slope' and 'y intercept', both represented as
'double.NaN'.