Implement geometric line-segment and ray
Add two new types using structs: LineSegment2 and Ray2.
Both are defined in terms of two dimensions.

Line is declared by passing two points.
Ray is declared by passing a single point and a direction (in radians).
If a given ray or line is parallel to a coordinate axis, the values of
the stored 'equation' variable will reflect it; as such, when using the
stored equation values in calculations, appropriate checks should be
implemented to prevent errors.