Calculate intersect of linear geometric objects
Implement a method to calculate linear intersection.
Take in an array of two 'Line2' line equations and use linear equations
to calculate the point of intersection. Pass through a series of
conditional statements to determine and use the appropriate equations
for calculation. Return a value representing the outcome and output the
result.