Add method for ray/line-segment intersection
Utilise the previously implemented "IntersectPoint" and "BetweenPoints"
methods together to calculate the point of intersection of a given 'ray'
and 'line' that is 'contained' between the two points that define the
'line'.

The method is capable of outputting the calculated intersection point,
as well as returning an 'int' that represents the 'status' of the
method.

The 'int' value returned, can be one of three values: 0, 1, or -1.
'0' means that no intersection was found. '1' means that a valid
intersection was calculated. '-1' means an error occurred. These values
should be used by the caller to detirmine an appropriate handling of the
result. If a valid point is found, the variable passed to the 'point'
parameter will be set to the resulting intersection point, due to the
use of the 'out' parameter modifier.

This method of returning a 'status' value and using the 'out' parameter
modifier, allows for simpler handling of the situations when no
intersection is found or when an error occurs. Instead of having to
validate the result, just check the return value of the method.