Calculate AB-line/turn-line intersection points
Obtain the points at which the current and target AB-lines intersect the
turn-line(s). Order the points based on distance. Group into pairs,
where the two points describe the bounds of a segement of AB-line.

Record the pair of points that bound the segment of AB-line that the
machine is currently on.