Update GUI.Designer.cs - fixed bug for UTurn mode
fixed bug: 
precondition: UTurn is on, and UTurn mode was set to the second mode (colored icon for UTurn)
then: going to the configuration menue and accept (green button). Settings are reloaded. UTurn is set to off; icon for UTurn mode has still the second mode icon (colored); but in AGO internaly UTurn mode is set to the advanced mode. When reactivating UTurn the UTurn Mode icon is vice versa.