Local search instead of global when following a curve
The previous implementation led the vehicle to sometimes pick a wrong part of a curve when the curve was intersecting itself (AB-curve) or there were more than one parallel lines near to each other (typically occurs when driving along an inner curve formed from a boundary curve and the field's shape is not convex).

This changes the curve-following algorithm to use a search for "local" nearest point instead of "global" (= taking all the points on the curve into account) when already following the curve.

Look ahead-distance is used when searching the nearest "local" point to prevent problems in tight hairpins caused by the search sticking to a local minimum "on the wrong side" of the hairpin (especially when using large look ahead values causing the vehicle to take a shortcut).

NOTE: Superseded PR was tested in a real tractor but this only on an external simulator. Behavior in the simulator seemed to be identical with the superseded PR. Implementation is not, however, completely identical fix due to the other changes in the source file, especially https://github.com/farmerbriantee/AgOpenGPS/pull/419

See the superseded PR (below) for more info.

Fixes https://github.com/farmerbriantee/AgOpenGPS/issues/391
Supersedes https://github.com/farmerbriantee/AgOpenGPS/pull/392